requirements.txt
absl-py==0.13.0
actionlib==1.12.1
angles==1.9.12
asn1crypto==0.24.0
astor==0.8.1
attrs==17.4.0
autobahn==17.10.1
Automat==0.6.0
backports.functools-lru-cache==1.4
backports.weakref==1.0.post1
bondpy==1.8.5
camera-calibration==1.15.0
camera-calibration-parsers==1.11.13
catkin==0.7.29
catkin-pkg==0.4.23
catkin-pkg-modules==0.4.23
catkin-tools==0.6.1
cbor==1.0.0
chardet==3.0.4
click==6.7
cloudpickle==1.1.1
colorama==0.3.7
constantly==15.1.0
contextlib2==0.6.0.post1
controller-manager==0.18.4
controller-manager-msgs==0.18.4
cryptography==2.1.4
cv-bridge==1.13.0
cycler==0.10.0
Cython==0.26.1
decorator==4.4.2
defusedxml==0.5.0
diagnostic-analysis==1.9.7
diagnostic-common-diagnostics==1.9.7
diagnostic-updater==1.9.7
dm-sonnet==1.36
docutils==0.14
dynamic-reconfigure==1.6.3
empy==3.3.2
enum34==1.1.10
funcsigs==1.0.2
futures==3.3.0
gast==0.5.2
gencpp==0.6.5
geneus==2.2.6
genlisp==0.4.16
genmsg==0.5.16
gennodejs==2.0.1
genpy==0.6.16
google-pasta==0.2.0
grpcio==1.39.0
h5py==2.10.0
hyperlink==17.3.1
idna==2.6
image-geometry==1.13.0
incremental==16.10.1
interactive-markers==1.11.5
ipaddress==1.0.17
joint-state-publisher==1.12.15
kdl-parser-py==1.13.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
keyring==10.6.0
keyrings.alt==3.0
laser-geometry==1.6.7
lz4==0.10.1
Markdown==3.1.1
matplotlib==2.1.1
message-filters==1.14.11
mock==3.0.5
mpi4py==2.0.0
netifaces==0.10.4
nose==1.3.7
numpy==1.16.6
olefile==0.45.1
osrf-pycommon==0.2.1
PAM==0.4.2
paramiko==2.0.0
pathlib==1.0.1
Pillow==5.1.0
protobuf==3.17.3
psutil==5.4.2
py-ubjson==0.8.5
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycairo==1.16.2
pycrypto==2.6.1
pycryptodomex==3.4.7
pydot==1.2.3
Pygments==2.2.0
pygobject==3.26.1
PyNaCl==1.1.2
PyOpenGL==3.1.0
pyOpenSSL==17.5.0
pyparsing==2.2.0
pyserial==3.4
python-dateutil==2.6.1
python-gnupg==0.4.1
python-qt-binding==0.4.4
python-snappy==0.5
PyTrie==0.2
pytz==2018.3
pyxdg==0.25
PyYAML==3.12
qrcode==5.3
qt-dotgraph==0.4.2
qt-gui==0.4.2
qt-gui-cpp==0.4.2
qt-gui-py-common==0.4.2
resource-retriever==1.12.6
roman==2.0.0
SecretStorage==2.3.1
semantic-version==2.8.5
sensor-msgs==1.12.8
service-identity==16.0.0
six==1.16.0
smach==2.0.1
smach-ros==2.0.1
smclib==1.8.5
subprocess32==3.2.7
tensorboard==1.14.0
tensorflow-estimator==1.14.0
tensorflow-probability==0.8.0
termcolor==1.1.0
tf==1.12.1
tf-conversions==1.12.1
tf2-geometry-msgs==0.6.5
tf2-kdl==0.6.5
tf2-py==0.6.5
tf2-ros==0.6.5
topic-tools==1.14.11
trollius==2.0.1
Twisted==17.9.0
txaio==2.8.1
u-msgpack-python==2.1
urdfdom-py==0.4.5
Werkzeug==1.0.1
wrapt==1.12.1
wsaccel==0.6.2
wxPython==3.0.2.0
wxPython-common==3.0.2.0
zope.interface==4.3.2
